Added Pokemon stat information. Made a PokemonType class. Included updated pokedex.db.
Modified database to include stat columns in the SQLite database, and
parsing of stats in PokeApiParser.
Made PokemonType class to extend off Pair, since using a Pair by itself
doesn't really make sense in terms of a meaningful design.
Created PokemonAbility class which will be used to represent Pokemon
ability (to be completed).
Included an updated pokedex.db to show the changes in the pokemon table.